requirements.txt
python-dotenv
mistralai
openpyxl